swap files from api to pages for backends run smoothly
