--- frontend/src/pages/api/settings.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export default function SettingsPage() {
-  return (
-    <div className="min-h-screen p-8 bg-gray-100">
-      <div className="max-w-3xl mx-auto bg-white p-6 rounded-xl shadow-md">
-        <h1 className="text-xl font-bold text-gray-800 mb-4">⚙️ Edit Profile</h1>
-        <p className="text-gray-600">Settings feature coming soon.</p>
-      </div>
-    </div>
-  );
-}
